Fixes #17439 - Moved facet base to a concern
If the plugin that wants to create a new facet doesn't want to
take the inheritance slot (for example if it already has some
base class), it can `include` the new concern.
